huggingface_space/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
torch==2.4.0
torchvision==0.19.0
Pillow==10.4.0
opencv-python-headless==4.10.0.84
facenet-pytorch==2.6.0
numpy==1.26.4
tqdm==4.66.5
pyyaml==6.0.2
